Add detection of already converted / transcribed assets

+++ utils/transcribe_audios.py
@@ -22,22 +22,23 @@ def transcribe_mp3_folder(
         filename for filename in os.listdir(input_folder) if filename.lower().endswith(".mp3")
     ]
     for filename in files_to_be_transcribed:
+
         input_path = os.path.join(input_folder, filename)
-        print(f"Transcribing {filename}...")
-
-        segments, info = pipeline.transcribe(input_path, beam_size=5)
-
-        output = {
-            "language": info.language,
-            "segments": [
-                {"start": seg.start, "end": seg.end, "text": seg.text}
-                for seg in segments
-            ]
-        }
         output_filename = os.path.splitext(filename)[0] + ".json"
         output_path = os.path.join(output_folder, output_filename)
 
-        with open(output_path, "w", encoding="utf-8") as f:
-            json.dump(output, f, indent=2)
-
-        print(f"Saved transcription to {output_path}")
+        if os.path.exists(output_path):
+            print(f"Transcription {output_path} already exists. Skipped. ")
+        else:
+            print(f"Transcribing {filename}...")
+            segments, info = pipeline.transcribe(input_path, beam_size=5)
+            output = {
+                "language": info.language,
+                "segments": [
+                    {"start": seg.start, "end": seg.end, "text": seg.text}
+                    for seg in segments
+                ]
+            }
+            with open(output_path, "w", encoding="utf-8") as f:
+                json.dump(output, f, indent=2)
+            print(f"Saved transcription to {output_path}")
